Rearrange the Rides form to a single component. Updated Create New Ride
